Cleaned up and tested
Tested this container locally and it ran fine.  I used this docker file to build it from a previously downloaded and unzipped copy of the core-nlp zip file.  This way the code is already in the container.  
I used this docker command to start it up.  You can drop the --restart unless-stopped.  I used it since it kept crashing and got tired of restarting it. Also, notice it's running on port 80 on my PC, but port 9000 inside the container. 
docker run -d -p 80:9000 --name core-nlp-service --restart unless-stopped elevenofjan/core-nlp-service:v1.3